requirements.txt
Django==3.2.25
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
requests==2.27.1
python-dotenv==0.19.2
django-cors-headers==3.10.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-apscheduler==0.6.2
tencentcloud-sdk-python==3.0.600
Pillow==8.4.0
Werkzeug==2.0.3
gunicorn==20.1.0
